Route 1a: add a Kotlin/Native Windows (mingwX64) target
Adds a mingwX64 native target that links skiko + Skia into a single Windows
DLL (skiko-windows-x64.dll) exporting the flat org_jetbrains_skia_* C surface,
generates a GNU import library (dumpbin -> .def -> dlltool), and points the
K/N cinterop linkerOpts at it. Only the extern-C surface crosses to K/N, so
the MSVC<->GNU C++ ABI wall is sidestepped and KT-65671 (fusing a C++ static
archive) is avoided. All new behaviour is opt-in via
-Pskiko.native.windows.enabled=true.

- sourceHierarchy.kt: group("windows"){ withMingwX64() } under native, so
  mingwX64 joins nativeMain + nativeJsMain (else the klib compiles empty).
- SkikoProjectContext: supportNativeWindows folded into supportAnyNative, so
  the @SymbolName opt-in (languageSettings.optIn) reaches the mingw compile.
- build.gradle.kts: configureNativeTarget(OS.Windows, Arch.X64, mingwX64())
  gated on the flag.
- NativeTasksConfiguration: OS.Windows compile branch (clang-cl +
  -DSKIKO_WINDOWS_DLL) and a new configureWindowsNativeTarget that links the
  DLL via LinkSkikoTask (lld-link, reusing the JVM Skia .libs), generates the
  import lib, and wires cinterop linkerOpts; force-exports the ICU uloc_*_skiko
  symbols that Kotlin binds via @SymbolName.
- common.h: SKIKO_EXPORT gains __declspec(dllexport) under SKIKO_WINDOWS_DLL.
- Resources.native.kt: ftell is 32-bit Int on Windows (LLP64), so size == 0L
  (Int vs Long) is normalised with .toLong().
- src/windowsMain: mingw actuals for the commonMain SkiaLayer /
  currentSystemTheme expects (copied from the linux stubs).

Build & publish (Windows host, VS Build Tools on SKIKO_VSBT_PATH):
  ./gradlew :skiko:publishKotlinMultiplatformPublicationToMavenLocal \
            :skiko:publishMingwX64PublicationToMavenLocal \
            -Pskiko.native.windows.enabled=true

+++ skiko/src/nativeMain/kotlin/org/jetbrains/skiko/Resources.native.kt
@@ -28,7 +28,9 @@ actual suspend fun loadBytesFromPath(path: String): ByteArray {
         throw Error("File '$path' is too long")
     }
 
-    if (size == 0L) {
+    // ftell returns C `long` = 32-bit Int on Windows (LLP64), 64-bit Long on
+    // Unix. `==` rejects mixed Int/Long (unlike </>), so normalise to Long.
+    if (size.toLong() == 0L) {
         fclose(file)
         return byteArrayOf()
     }
